BLEH_application/requirements.txt
OpenAI
fastapi 
uvicorn[standard]
gunicorn
python-dotenv